+++ aws/gen_api.go
@@ -67,6 +67,7 @@ var ResourceTypes = []string{
 	"availabilityzone",
 	"loadbalancer",
 	"targetgroup",
+	"listener",
 	"user",
 	"group",
 	"role",
@@ -99,6 +100,7 @@ var ServicePerResourceType = map[string]string{
 	"availabilityzone": "infra",
 	"loadbalancer":     "infra",
 	"targetgroup":      "infra",
+	"listener":         "infra",
 	"user":             "access",
 	"group":            "access",
 	"role":             "access",
@@ -149,6 +151,7 @@ func (s *Infra) ResourceTypes() (all []string) {
 	all = append(all, "availabilityzone")
 	all = append(all, "loadbalancer")
 	all = append(all, "targetgroup")
+	all = append(all, "listener")
 	return
 }
 
@@ -167,6 +170,7 @@ func (s *Infra) FetchResources() (*graph.Graph, error) {
 	var availabilityzoneList []*ec2.AvailabilityZone
 	var loadbalancerList []*elbv2.LoadBalancer
 	var targetgroupList []*elbv2.TargetGroup
+	var listenerList []*elbv2.Listener
 
 	errc := make(chan error)
 	var wg sync.WaitGroup
@@ -296,6 +300,18 @@ func (s *Infra) FetchResources() (*graph.Graph, error) {
 		var resGraph *graph.Graph
 		var err error
 		resGraph, targetgroupList, err = s.fetch_all_targetgroup_graph()
+		if err != nil {
+			errc <- err
+			return
+		}
+		g.AddGraph(resGraph)
+	}()
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		var resGraph *graph.Graph
+		var err error
+		resGraph, listenerList, err = s.fetch_all_listener_graph()
 		if err != nil {
 			errc <- err
 			return
@@ -468,6 +484,19 @@ func (s *Infra) FetchResources() (*graph.Graph, error) {
 			}
 		}
 	}()
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		for _, r := range listenerList {
+			for _, fn := range addParentsFns["listener"] {
+				err := fn(g, r)
+				if err != nil {
+					errc <- err
+					return
+				}
+			}
+		}
+	}()
 
 	go func() {
 		wg.Wait()
@@ -517,6 +546,9 @@ func (s *Infra) FetchByType(t string) (*graph.Graph, error) {
 		return graph, err
 	case "targetgroup":
 		graph, _, err := s.fetch_all_targetgroup_graph()
+		return graph, err
+	case "listener":
+		graph, _, err := s.fetch_all_listener_graph()
 		return graph, err
 	default:
 		return nil, fmt.Errorf("aws infra: unsupported fetch for type %s", t)