+++ gen/aws/fetchers_definitions.go
@@ -50,6 +50,7 @@ var FetchersDefs = []fetchersDef{
 			{ResourceType: graph.AvailabilityZone.String(), AWSType: "ec2.AvailabilityZone", ApiMethod: "DescribeAvailabilityZones", Input: "ec2.DescribeAvailabilityZonesInput{}", Output: "ec2.DescribeAvailabilityZonesOutput", OutputsExtractor: "AvailabilityZones"},
 			{ResourceType: graph.LoadBalancer.String(), AWSType: "elbv2.LoadBalancer", ApiMethod: "DescribeLoadBalancersPages", Input: "elbv2.DescribeLoadBalancersInput{}", Output: "elbv2.DescribeLoadBalancersOutput", OutputsExtractor: "LoadBalancers", Multipage: true, NextPageMarker: "NextMarker"},
 			{ResourceType: graph.TargetGroup.String(), AWSType: "elbv2.TargetGroup", ApiMethod: "DescribeTargetGroups", Input: "elbv2.DescribeTargetGroupsInput{}", Output: "elbv2.DescribeTargetGroupsOutput", OutputsExtractor: "TargetGroups"},
+			{ResourceType: graph.Listener.String(), AWSType: "elbv2.Listener", ManualFetcher: true},
 		},
 	},
 	{